+++ src/sovereign/utils/mock.py
@@ -14,10 +14,10 @@ def mock_discovery_request(service_cluster=None, resource_names=None, region='no
                 zone=region
             ),
             metadata={
-                'hide_private_keys': True,
                 **metadata
             }
         ),
         version_info=str(randint(100000, 1000000000)),
         resource_names=resource_names or [],
+        hide_private_keys=True,
     )